# alzuhud.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "alzuhud services",
    "version": "1.1.0"
  },
  "host": "165.22.234.156:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/tickets/sections/{ticketId}": {
      "get": {
        "operationId": "ticketSections",
        "responses": {
          "200": {
            "description": "ticket sections",
            "schema": {
              "$ref": "#/definitions/AlzuhudResponseTicketSections"
            }
          }
        },
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "tags": [
          "Alzuhud"
        ]
      }
    },
    "/api/guest/contact/add": {
      "post": {
        "operationId": "contact",
        "responses": {
          "200": {
            "description": "contact",
            "schema": {
              "$ref": "#/definitions/AlzuhudResponseVoid"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AlzuhudRequestAddContact"
            }
          }
        ],
        "tags": [
          "Alzuhud"
        ]
      }
    },
...
  },
  "definitions": {
...
    "AlzuhudRequestAddContact": {
      "type": "object",
      "properties": {
        "subject": {
          "type": "string",
          "description": "required and max length is 50"
        },
        "message": {
          "type": "string",
          "description": "required and max length is 150000"
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "optional and max length is 200 and must be a valid email address"
        },
        "name": {
          "type": "string",
          "description": "optional and max length is 50"
        }
      }
    },
...
    "AlzuhudResponseVoid": {
      "type": "object"
    },
...
    "UserObjectDefintion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "full_name": {
          "type": "string",
          "format": "string"
        },
        "email": {
          "type": "string",
          "format": "string"
        },
        "phone": {
          "type": "string",
          "format": "string"
        },
        "role": {
          "type": "string",
          "format": "string"
        },
        "status": {
          "type": "string",
          "format": "string"
        },
        "email_verified_at": {
          "type": "string",
          "format": "timestamp"
        },
        "remember_token": {
          "type": "string",
          "format": "timestamp"
        },
        "created_at": {
          "type": "string",
          "format": "string"
        },
        "updated_at": {
          "type": "string",
          "format": "timestamp"
        }
      }
    },
...
    "ReferenceAllTicketSectionsArray": {
      "type": "object",
      "properties": {
        "ticket_id": {
          "type": "integer",
          "format": "int64"
        },
        "user_id": {
          "$ref": "#/definitions/UserObjectDefintion"
        },
        "body": {
          "type": "string"
        },
        "user_role": {
          "type": "string"
        },
        "ticket_img": {
          "type": "string"
        },
        "ticket_img_url": {
          "type": "string"
        }
      }
    },
    "AlzuhudResponseTicketSections": {
      "type": "object",
      "properties": {
        "ticket": {
          "type": "object",
          "properties": {
            "user_id": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/UserObjectDefintion"
              }
            },
            "subject": {
              "type": "string"
            },
            "body": {
              "type": "string"
            },
            "status": {
              "type": "string"
            },
            "ticket_img": {
              "type": "string"
            },
            "ticket_img_url": {
              "type": "string"
            }
          }
        },
        "sections": {
          "type": "object",
          "properties": {
            "current_page": {
              "type": "integer",
              "format": "int64"
            },
            "data": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReferenceAllTicketSectionsArray"
              }
            },
            "first_page_url": {
              "type": "string",
              "format": "string"
            },
            "from": {
              "type": "integer",
              "format": "int64"
            },
            "last_page": {
              "type": "integer",
              "format": "int64"
            },
            "last_page_url": {
              "type": "string",
              "format": "url"
            },
            "next_page_url": {
              "type": "string",
              "format": "url"
            },
            "path": {
              "type": "string",
              "format": "uri"
            },
            "per_page": {
              "type": "integer",
              "format": "int32"
            },
            "prev_page_url": {
              "type": "integer",
              "format": "int32"
            },
            "to": {
              "type": "integer",
              "format": "int64"
            },
            "total": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      }
    },
...
  }
}